refactor(dashboard): extract grid geometry helpers; delete legacy useDashboardLayout
useCatalogLayout depended on compactVertically/swapOnDrop from the 641-line legacy
useDashboardLayout hook (the only thing still importing it). Move those two pure
helpers (+ their rectsOverlap/overlapArea/collidesAt deps) into utils/gridGeometry.js
and delete useDashboardLayout.js entirely. This also drops the last live importer of
layoutConfig's WIDGETS/DEFAULT_LAYOUT (trimmed next).
